Make EMI + Jade plugins cross-platform
compiling against Jade's Fabric API in the common module should be *fine*
