+++ fabric/build.gradle
@@ -19,10 +19,6 @@ configurations {
 	developmentFabric.extendsFrom common
 }
 
-repositories {
-	maven { url "https://maven.terraformersmc.com" }
-}
-
 dependencies {
 	modImplementation "net.fabricmc:fabric-loader:$fabric_loader_version"
 	modApi "net.fabricmc.fabric-api:fabric-api:$fabric_version"
@@ -35,8 +31,6 @@ dependencies {
 //	modCompileOnly "mezz.jei:jei-$minecraft_version-fabric-api:$jei_version"
 //	modRuntimeOnly "mezz.jei:jei-$minecraft_version-fabric:$jei_version"
 
-	modCompileOnly "dev.emi:emi-xplat-intermediary:${emi_version}:api"
-	modCompileOnly "dev.emi:emi-fabric:${emi_version}:api"
 	modLocalRuntime "dev.emi:emi-fabric:${emi_version}"
 
 	modRuntimeOnly "curse.maven:jade-324717:4978962"  // Jade-1.20.4-fabric-13.2.1
